jieli: bound UART admission and retain in-flight DMA storage
Nonblocking console reads waited for writers, and dev_write could pend its
completion semaphore forever after admission consumed the caller budget.
Use the fixed SDK non-IRQ policy, one elapsed deadline and owned staging
retained until TX completion, including after partial/zero-wait writes.

Six real-source fault scenarios fail before; seven pass after with strict
Clang/GCC and pthread/TSan coverage. The five original UART fixtures pass.
Native Loader, PAL and display builds, host-only iOS analysis, guides and
diff checks pass. The candidate App passes PAL Filesystem 13, Core and
Wi-Fi 27 (10/10). Pre-decoder capture proves earlier omitted monitor lines
were intact on UART. Final Loader/lifecycle acceptance remains pending;
USB and KCP deadline work remains open.

+++ projects/example/native_component_src/jieli/wl82/devkit_app/src/color_bar_pal.c
@@ -303,7 +303,7 @@ static int draw_color_bars(const h2_pal_display_api_t *display) {
   if (info.width == 0 || info.width > H2_LCD_MAX_WIDTH || info.height == 0) {
     return H2_DISPLAY_ERR_INVALID_ARG;
   }
-  for (uint32_t x = 0u; x < info.width; ++x) {
+  for (uint32_t x = 0u; x < (uint32_t)info.width; ++x) {
     color_line[x] = colors[x * 8u / info.width];
   }
   for (int y = 0; y < info.height; ++y) {
